LayoutTests/http/wpt/identity/identitycredentialscontainer-get-basics.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Digital Credential API: get() default behavior checks.</title>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body></body>
<script>
    promise_setup(async () => {
        if (document.visibilityState === "hidden") {
            await new Promise((resolve) => {
                document.onvisibilitychange = resolve;
                testRunner.setPageVisibility("visible");
            });
        }
        assert_equals(document.visibilityState, "visible", "should be visible");
    });

    promise_test(async (t) => {
        await test_driver.bless();
        await promise_rejects_dom(
            t,
            "NotSupportedError",
            navigator.identity.get(),
            "navigator.identity.get() with no argument."
        );

        await test_driver.bless();
        await promise_rejects_dom(
            t,
            "NotSupportedError",
            navigator.identity.get({}),
            "navigator.identity.get() with empty dictionary."
        );

        await test_driver.bless();
        await promise_rejects_js(
            t,
            TypeError,
            navigator.identity.get({ digital: "wrong type" }),
            "navigator.identity.get() with bogus digital type"
        );

        await test_driver.bless();
        await promise_rejects_dom(
            t,
            "NotSupportedError",
            navigator.identity.get({ bogus_key: "bogus" }),
            "navigator.identity.get() with unknown key (same as passing empty dictionary)."
        );

        await test_driver.bless();
        await promise_rejects_js(
            t,
            TypeError,
            navigator.identity.get({ digital: { providers: [] } }),
            "navigator.identity.get() with an empty list of providers"
        );

        await test_driver.bless();
        await promise_rejects_js(
            t,
            TypeError,
            navigator.identity.get({
                digital: { providers: [{ protocol: "bogus protocol", request: {} }] },
            }),
            "navigator.identity.get() with a provider with unknown protocol"
        );

        const controller = new AbortController();
        const options = { signal: controller.signal };

        controller.abort();
        await promise_rejects_dom(
            t,
            "AbortError",
            navigator.identity.get(options),
            "navigator.identity.get() with abort signal set"
        );

        assert_equals(
            await navigator.identity.get({
                digital: { providers: [{ protocol: "mdoc", request: {} }] },
            }),
            null,
            "navigator.identity.get() with a valid provider"
        );
    }, "navigator.identity.get() default behavior checks.");
</script>
